separate out script from swing for FG/MergeHM CLI
Very quick separation of JFrame elements from script elements into
respective scripts/* and window_interface/*/*Output classes. The script
worke well as a public static void method so I went ahead and made it a
static method.

CLI class has minimal validation as the inputs are a sense.png,
anti.png, and output.png. The comments within CLI outline possible
improvements to make the user input more robust. These are added to the
todo list and will come back to this when a discussion is had about
whether to allow for input of only one heatmap to output to the same
heatmap (for generalized pipeline purposes). Also possible checks on
filenames for "_sense" and "_anti" may be useful.

The jar executes MergeHeatMap without errors on ACI.

cli/Figure_Generation/HeatMapCLI class added a line to indicate
completion of image generation (mimic MergeHeatMap).
