+++ src/cli/Figure_Generation/MergeHeatMapCLI.java
@@ -10,6 +10,8 @@ import java.util.List;
 
 import java.io.File;
 
+import scripts.Figure_Generation.MergeHeatMapPlot;
+
 /**
 	Figure_GenerationCLI/MergeHeatMapCLI
 */
@@ -17,16 +19,47 @@ import java.io.File;
 		description = "Merge Sense and Antisense png heatmaps")
 public class MergeHeatMapCLI implements Callable<Integer> {
 	
-	@Parameters( index = "0", arity = "1..3", description = "list of PNG files")
-	private List<File> inputFile = null;
 	
-	@Option(names = {"-o", "--output"}, description = "specify output file ")
-	private File output = new File("output.matrix");
+	@Option(names = {"-s", "--sense"}, required=true, description = "First(sense) heatmap to merge")
+	private File senseFile = null;
+	@Option(names = {"-a", "--anti"}, description = "Second(anti) heatmap to merge")
+	private File antiFile = null;
+	
+	@Option(names = {"-o", "--output"}, description = "specify output file (be sure to use .png ext)")
+	private File output = new File("merged.png");
 	
 	@Override
 	public Integer call() throws Exception {
 		System.out.println( ">MergeHeatMapCLI.call()" );
-// 		SEStats stat = new SEStats( bamFile, output );		
+		
+		if( validateInput()!=0 ){
+			System.err.println("Invalid input. Check usage using '-h' or '--help'");
+			return(1);
+		}
+		
+		// Generate Merged HeatMap
+		MergeHeatMapPlot.mergePNG( senseFile, antiFile, output );
+				
+		System.out.println( "Image Generated." );
+			
+		return(0);
+	}
+	
+	private int validateInput(){
+		
+		if( senseFile==null && antiFile==null ){
+			System.err.println("!!!What are we merging? You gave two null file inputs!");
+			return(1);
+		}
+		
+		//check input extensions
+		
+		//check outputbasename is valid
+		
+		//Print warnings if not named _anti and _sense
+		
+		//Possibly print warnings when image has color values that dont match red or blue?
+		
 		return(0);
 	}
